+++ src/main/java/courgette/api/MobileDeviceType.java
@@ -1,0 +1,7 @@
+package courgette.api;
+
+public enum MobileDeviceType {
+    SIMULATOR,
+    REAL_DEVICE,
+    SIMULATOR_AND_REAL_DEVICE
+}