+++ src/main/java/courgette/runtime/CourgetteRunnerInfo.java
@@ -10,16 +10,20 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 public class CourgetteRunnerInfo {
     private final CourgetteRuntimeOptions courgetteRuntimeOptions;
+    private final CourgetteProperties courgetteProperties;
     private final Integer lineId;
     private final CourgetteRunLevel courgetteRunLevel;
     private final Feature feature;
     private final boolean rerun;
+    private final DeviceType deviceType;
 
     public CourgetteRunnerInfo(CourgetteProperties courgetteProperties, Feature feature, Integer lineId) {
+        this.courgetteProperties = courgetteProperties;
         this.feature = feature;
         this.courgetteRuntimeOptions = new CourgetteRuntimeOptions(courgetteProperties, feature);
         this.lineId = lineId;
@@ -28,6 +32,7 @@ public class CourgetteRunnerInfo {
                 checkRerunCondition(feature,
                         courgetteProperties.getCourgetteOptions().excludeFeatureFromRerun(),
                         courgetteProperties.getCourgetteOptions().excludeTagFromRerun());
+        this.deviceType = determineDeviceType().orElse(null);
     }
 
     public Feature getFeature() {
@@ -83,6 +88,47 @@ public class CourgetteRunnerInfo {
 
     public boolean allowRerun() {
         return rerun;
+    }
+
+    public DeviceType getDeviceType() {
+        return deviceType;
+    }
+
+    private Optional<DeviceType> determineDeviceType() {
+        if (courgetteProperties.isMobileDeviceAllocationPluginEnabled()) {
+            switch (courgetteProperties.getCourgetteOptions().mobileDeviceType()) {
+                case SIMULATOR:
+                    return Optional.of(DeviceType.SIMULATOR);
+                case REAL_DEVICE:
+                    return Optional.of(DeviceType.REAL_DEVICE);
+                case SIMULATOR_AND_REAL_DEVICE:
+                    if (courgetteRunLevel.equals(CourgetteRunLevel.FEATURE)) {
+                        if (feature.getPickles().stream()
+                                .anyMatch(pickle -> pickle.getTags().stream()
+                                        .map(String::toLowerCase)
+                                        .anyMatch(tag ->
+                                                Arrays.stream(courgetteProperties.getCourgetteOptions().realMobileDeviceTag())
+                                                        .map(String::toLowerCase)
+                                                        .anyMatch(realDeviceTag -> realDeviceTag.contains(tag))
+                                        ))) {
+                            return Optional.of(DeviceType.REAL_DEVICE);
+                        }
+                    } else {
+                        if (feature.getPickles().stream()
+                                .filter(pickle -> pickle.getLocation().getLine() == lineId)
+                                .anyMatch(pickle -> pickle.getTags().stream()
+                                        .map(String::toLowerCase)
+                                        .anyMatch(tag ->
+                                                Arrays.stream(courgetteProperties.getCourgetteOptions().realMobileDeviceTag())
+                                                        .map(String::toLowerCase)
+                                                        .anyMatch(realDeviceTag -> realDeviceTag.equals(tag))
+                                        ))) {
+                            return Optional.of(DeviceType.REAL_DEVICE);
+                        }
+                    }
+            }
+        }
+        return Optional.of(DeviceType.SIMULATOR);
     }
 
     private boolean checkRerunCondition(Feature feature, String[] excludedRerunFeatures, String[] excludedRerunTags) {