+++ src/main/java/courgette/runtime/DeviceType.java
@@ -1,0 +1,6 @@
+package courgette.runtime;
+
+public enum DeviceType {
+    SIMULATOR,
+    REAL_DEVICE
+}