+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+CHANGES IN VERSION 6.11.0
+=================================
+* [NEW] Add new Courgette option `mobileDeviceType` to support running tests that requires simulators and real devices in the same run.
+* [NEW] Add new Courgette option `realMobileDeviceTag` to allow Courgette to allocate real devices to tests with a matching tag.
+* [MAINTENANCE] Upgrade to Cucumber version 7.15.0
+
 CHANGES IN VERSION 6.10.0
 =================================
 * [MAINTENANCE] Upgrade to Cucumber version 7.12.1